// src/app/features/executions/types.ts
import { type Realtime } from "@inngest/realtime";
import { type GetStepTools, type Inngest } from "inngest";

export type WorkflowContext = Record<string, unknown>;


export type StepTools = GetStepTools<Inngest.Any>;

export interface NodeExecutorParams<Tdata=Record<string, unknown>> {
    data: Tdata;
    nodeId: string;
    userId: string;
    context: WorkflowContext;
    step: StepTools;
    publish: Realtime.PublishFn;
}

export type NodeExecutor<Tdata=Record<string, unknown>> = (params: NodeExecutorParams<Tdata>) => Promise<WorkflowContext>;